% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{geopr_list_layers}
\alias{geopr_list_layers}
\title{Listar capas disponibles del WFS de Puerto Rico (GeoServer PR) con metadatos}
\usage{
geopr_list_layers(
  wfs_url = "http://geoserver2.pr.gov/geoserver/pr_geodata/wfs",
  serviceVersion = "2.0.0",
  use_cache = TRUE
)
}
\arguments{
\item{wfs_url}{URL del servicio WFS.}

\item{serviceVersion}{Versión WFS.}

\item{use_cache}{Si TRUE, reutiliza el resultado en memoria durante la sesión.}
}
\description{
Listar capas disponibles del WFS de Puerto Rico (GeoServer PR) con metadatos
}
